+++ packages/tools/prettier/config.cjs
@@ -1,0 +1,2 @@
+﻿// @ts-ignore
+module.exports = require('./config.json');