+++ packages/tools/prettier/config.mjs
@@ -1,0 +1,2 @@
+﻿// @ts-ignore
+export * from './config.cjs';