+++ packages/tools/rollup/helpers.mjs
@@ -1,0 +1,4 @@
+﻿// @ts-ignore
+import helpers from './helpers.cjs';
+
+export const {escape, dump, Scopes, Resolver, Locator} = helpers;