+++ packages/tools/CHANGELOG.md
@@ -1,14 +1,31 @@
 ﻿# Change Log
 
-## [0.0.2][]
+## [0.0.3][] — `unreleased`
+
+> **Release** — `@smotaal.io/tools@0.0.2`
 
 ### Added
 
-- Refactored `@smotaal.io/tools/prettier.config` from `@smotaal.io/prettier.config`.
+- New `rollup` configuration & helpers — [changelog](./rollup/CHANGELOG.md#0.0.1).
 
-## [0.0.1][]
+### Changed
+
+- Renamed `prettier.config` to `prettier` (still configuration only) — [changelog](./prettier/CHANGELOG.md#0.0.2).
+
+## [0.0.2][] - 2019-11-22
+
+> **Release** — `@smotaal.io/tools@0.0.2`
+
+### Added
+
+- Refactored `prettier.config` configuration — [changelog](./prettier/CHANGELOG.md#0.0.1).
+
+## [0.0.1][] — 2019-11-15
+
+> **Release** — `@smotaal.io/tools@0.0.1`
 
 - Initial [`README.md`](./README.md) and `package.json`.
 
+[0.0.3]: https://www.npmjs.com/package/@smotaal.io/tools/v/0.0.3
 [0.0.2]: https://www.npmjs.com/package/@smotaal.io/tools/v/0.0.2
 [0.0.1]: https://www.npmjs.com/package/@smotaal.io/tools/v/0.0.1